requirements.txt
requests
python-dotenv
ollama
fast-flights
